[COREVM-118] Move bytecode_loader into public namespace
